+++ R/convertPaths.R
@@ -130,6 +130,7 @@ setMethod(
     } else {
       fns <- NULL
     }
+    # DON"T BE TEMPTED TO RM NZ i.e., MEMORY LAYERS -- NEED TO KNOW THAT THERE WERE SOME
     normPath(fns)
 })